Try to fix Travis tests
Issue #207
